Script: common base class for all write scripts
