+++ server/src/main/java/org/elasticsearch/script/UpdateScript.java
@@ -14,7 +14,7 @@ import java.util.Map;
 /**
  * A script used in the update API
  */
-public abstract class UpdateScript {
+public abstract class UpdateScript extends WriteScript {
 
     public static final String[] PARAMETERS = {};
 
@@ -24,26 +24,14 @@ public abstract class UpdateScript {
     /** The generic runtime parameters for the script. */
     private final Map<String, Object> params;
 
-    private final UpdateCtxMap ctxMap;
-
     public UpdateScript(Map<String, Object> params, UpdateCtxMap ctxMap) {
+        super(ctxMap);
         this.params = params;
-        this.ctxMap = ctxMap;
     }
 
     /** Return the parameters for this script. */
     public Map<String, Object> getParams() {
         return params;
-    }
-
-    /** Return the update context for this script. */
-    public Map<String, Object> getCtx() {
-        return ctxMap;
-    }
-
-    /** Return the update metadata for this script */
-    public Metadata metadata() {
-        return ctxMap.getMetadata();
     }
 
     public abstract void execute();
